(feat) add mTLS and ip network exclusion support
mTLS support adds the ability to check of a client provided TLS
certificate was signed by the configured ca. ip net explusion
supports makes it possible to bypass authentication requirements
for requests originating from specific "exclusion" networks.

this commit changes a few configuration options and names which
will be updated in the documentation in a following commit.